POW fix
